feat: configure tests to run in both browser and node env
