+++ packages/accounts/package.json
@@ -34,7 +34,7 @@
         "bs58": "4.0.0",
         "jest": "26.0.1",
         "near-hello": "0.5.1",
-        "near-workspaces": "3.4.0",
+        "near-workspaces": "3.5.0",
         "ts-jest": "26.5.6",
         "typescript": "4.9.4"
     },
