+++ packages/crypto/package.json
@@ -20,7 +20,8 @@
     "@near-js/utils": "workspace:*",
     "borsh": "1.0.0",
     "@noble/curves": "1.2.0",
-    "randombytes": "2.1.0"
+    "randombytes": "2.1.0",
+    "patch-package": "8.0.0"
   },
   "devDependencies": {
     "@types/node": "18.11.18",
