+++ packages/keystores/test/in_memory_keystore.test.js
@@ -7,5 +7,5 @@ describe('In-memory keystore', () => {
         ctx.keyStore = new InMemoryKeyStore();
     });
 
-    require('./keystore_common').shouldStoreAndRetriveKeys(ctx);
+    require('./keystore_common').shouldStoreAndRetrieveKeys(ctx);
 });
