+++ packages/keystores/test/keystore_common.js
@@ -4,7 +4,7 @@ const NETWORK_ID_SINGLE_KEY = 'singlekeynetworkid';
 const ACCOUNT_ID_SINGLE_KEY = 'singlekey_accountid';
 const KEYPAIR_SINGLE_KEY = new KeyPairEd25519('2wyRcSwSuHtRVmkMCGjPwnzZmQLeXLzLLyED1NDMt4BjnKgQL6tF85yBx6Jr26D2dUNeC716RBoTxntVHsegogYw');
 
-module.exports.shouldStoreAndRetriveKeys = ctx => {
+module.exports.shouldStoreAndRetrieveKeys = ctx => {
     beforeEach(async () => {
         await ctx.keyStore.clear();
         await ctx.keyStore.setKey(NETWORK_ID_SINGLE_KEY, ACCOUNT_ID_SINGLE_KEY, KEYPAIR_SINGLE_KEY);
