+++ packages/providers/package.json
@@ -25,7 +25,7 @@
   "devDependencies": {
     "@types/node": "18.11.18",
     "jest": "26.0.1",
-    "near-workspaces": "3.4.0",
+    "near-workspaces": "3.5.0",
     "ts-jest": "26.5.6",
     "typescript": "4.9.4"
   },
