+++ packages/signers/test/signer.test.js
@@ -2,6 +2,9 @@ const { InMemoryKeyStore } = require('@near-js/keystores');
 
 const { InMemorySigner } = require('../lib');
 
+const { TextEncoder } = require('util');
+global.TextEncoder = TextEncoder;
+
 test('test no key', async() => {
     const signer = new InMemorySigner(new InMemoryKeyStore());
     await expect(signer.signMessage('message', 'user', 'network'))
